Allow "black lies" to be cached
"black lies" differ from "white lies" in that the owner name
of the NSEC record matches the QNAME and the intent is to
return NODATA instead of NXDOMAIN for all types.  Caching
this NSEC does not lead to unexpected behaviour on synthesis
when the QNAME matches the NSEC owner which it does for the
the general "white lie" response.